Expose releases as an api
